# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Antipa API Documentation",
        "description": "Documentation de l'API Antipa",
        "version": "1.0.0"
    },
    "paths": {
        "/api/diseases": {
            "get": {
                "summary": "Récupérer la liste des maladies",
                "operationId": "1d8c114c95b03eaae97cd6a1e07fa9e4",
                "responses": {
                    "200": {
                        "description": "Liste des maladies"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de l'affichage des maladies."
                    }
                }
            },
            "post": {
                "summary": "Créer une nouvelle maladie",
                "operationId": "31843b1cf4f9f400ab9b5f6529ebb22a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Maladie créée avec succès"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la création de la maladie."
                    }
                }
            }
        },
        "/api/diseases/{id}": {
            "get": {
                "summary": "Afficher une maladie spécifique",
                "operationId": "a87b653683ba9a7f4417723db7d1214f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la maladie"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de l'affichage de la maladie."
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour une maladie spécifique",
                "operationId": "dce80ce3af88550396f356e7f594b15b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Maladie mise à jour avec succès"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la mise à jour de la maladie."
                    }
                }
            },
            "delete": {
                "summary": "Supprimer une maladie spécifique",
                "operationId": "f7caca721967626c8b5b7021251f7ee4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Maladie supprimée avec succès"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la suppression de la maladie."
                    }
                }
            }
        },
        "/api/localizations": {
            "get": {
                "summary": "Récupérer la liste des localisations",
                "operationId": "ca9d99636d9a01f672483931dc6d6a29",
                "responses": {
                    "200": {
                        "description": "Liste des localisations"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de l'affichage des localisations."
                    }
                }
            },
            "post": {
                "summary": "Créer une nouvelle localisation",
                "operationId": "9bf538f408cd2b41a6887a8ee83170f2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "type": "string"
                                    },
                                    "continent": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Localisation créée avec succès"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la création de la localisation."
                    }
                }
            }
        },
        "/api/localizations/{id}": {
            "get": {
                "summary": "Afficher une localisation spécifique",
                "operationId": "78c32ba8ad26b8fa723661a717e10fa5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails de la localisation"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de l'affichage de la localisation."
                    }
                }
            },
            "put": {
                "summary": "Mettre à jour une localisation spécifique",
                "operationId": "f67cc9330f64e3d718c8d974c7b05e65",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "country": {
                                        "type": "string"
                                    },
                                    "continent": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Localisation mise à jour avec succès"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la mise à jour de la localisation."
                    }
                }
            },
            "delete": {
                "summary": "Supprimer une localisation spécifique",
                "operationId": "fbb4f8aead127c93368b29b0d1094b61",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Localisation supprimée avec succès"
                    },
                    "500": {
                        "description": "Une erreur est survenue lors de la suppression de la localisation."
                    }
                }
            }
        },
...
    }
}